requirements.txt
# ----------------------------
# Core LLM & Agentic Framework
# ----------------------------
langchain==0.3.26
langchain-community==0.3.27
langchain-chroma==0.2.5
langchain-huggingface==0.0.3


# Embeddings + Transformers
sentence-transformers==2.7.0
huggingface_hub==0.24.5

# ----------------------------
# Vector Store & Embeddings
# ----------------------------
chromadb==1.0.9
faiss-cpu==1.10.0

# ----------------------------
# PDF & Document Processing
# ----------------------------
pypdf==4.3.1
pdfplumber==0.11.4
unstructured==0.15.3

# ----------------------------
# Web UI & API
# ----------------------------
streamlit==1.37.1
fastapi==0.115.9
uvicorn==0.30.3

# ----------------------------
# Data & Scientific Stack
# ----------------------------
numpy==1.26.4
pandas==2.2.2

# ----------------------------
# Deep Learning / Vision
# ----------------------------
torch==2.2.2
torchvision==0.17.2
tensorflow==2.18.0
ultralytics==8.1.34
opencv-python==4.9.0.80

# ----------------------------
# Utilities & Logging
# ----------------------------
tqdm==4.66.4
python-dotenv==1.0.1
requests==2.32.3
